Add "mini" query page tab (#1487)
Add a new 'mini' query page tab which displays a mini version of the
query page in a tab on the viewer page. This is useful for quickly
modifying queries and seeing the results, while being able to compare
results against the timeline.

This tab is behind a flag initially as it's very rough and still needs
some polish.

Going forward, this tab should sync with the query page itself, taking
the place of the 'Standalone Query' tab that's created on the viewer
page whenever a query is run on the query page.