Improve wasm64 support (#1492)
- Reduce duplication in .gni files
- Make 64-bit detection synchronous. The async adds extra complexity
that can be avoided.
- Change the way 64-bit-only is passed. I am not a an of the "async
import" as that
IMHO creates unnecessary risks and makes the code even more complicate
due to
an unnecessary async. Instad use a stub import + rollup to redirect to
the real import.
- Rollback some of the changes to engine.ts. I don't like the extra
layers introduced by
#1458. I think the previous model where wasm_bridge.ts does everything
is cleaner.
- Fix a bug which #1458 accidentally re-introduced on wasm32, where
addresses that are
 between 2GB and 4GB accidentally become negative offsets.

I tested it both on Chrome (which supports wasm64) and an older Firefox
(which doesn't)