Add plugin to handle entity state residency aggregation
Add a plugin to handle entity state residency. This plugin aims to
present existing aggregation data in a more user-friendly manner,
as well a remove aggregations that do not make sense. This is similar
in concept to https://github.com/google/perfetto/pull/1211

This commit DOES NOT aim to fix any existing errors in computation
like is mentioned in https://github.com/google/perfetto/issues/1478
and instead just aims to focus on better labeling for existing data

BUG: https://github.com/google/perfetto/issues/1321