frozen_ftrace: Erase the trace data after read
After read out all data and flush the writer from
"linux.frozen_ftrace" data source, make sure to clear trace data
in the kernel.

Background of this change is the upstream kernel patch[1], which
makes the kernel to recover the trace data, which is already read
in the previous boot. This actually expects the perfetto AOT use
case, because it reads out the kernel data periodically. Thus we
can not read most of the data in the persistent ring buffer next
boot time.
However, even if perfetto read out the previous boot data via
"linux.frozen_ftrace" data source, it recovers the data too
(unless it is reused.) This means perfetto will read the same
data again.

This does not change anything from user's viewpoint, because on
the current kernel, the read data is gone. And on the kernel with
the patch[1], it still be gone.

[1] https://lore.kernel.org/all/174709742769.1964496.18203163435305117893.stgit@mhiramat.tok.corp.google.com/

Change-Id: Id4a7852c493fc4ac635396314eceb9d52ac57732