Android Java SDK: don't call native methods in the 'Category' constructor. (#1481)
If the Category is stored as a static final field, its constructor is
called on the class loading phase. That requires the native library also
to be loaded in the class's static initialization block, which is not
what we usually want.

In this commit, we also fix the race condition between `#register()`
and `#unregister()` methods.

To test the new initialization workflow, the `System.loadLibrary` call
in `PerfettoTraceTest` class is moved from the `static {}` block to the
`setUp` method. This method is called at runtime after all static final
fields are created.

Tested:
```
tools/bazel test //:src_android_sdk_java_test_perfetto_trace_instrumentation_test
```