Disable some heapprofd cts tests for native bridge [cherry-pick] (#1506)
These tests are relying on the fact that callstacks can provide
information about which function called malloc/free.

This is not the case for apps running with native_bridge. For these
there exists 2 different stacks (native one - visible to perfetto) and
another one for emulated architecture (to access which perfetto needs to
read and unwind second, - guest stack).

The result is that as of now all native callstacks for memory
allocations originated for emulated apps look the same and get
aggregated together. And this is why test does not pass.

Bug: 411111586
Test: run heapprofd tests on Android Emulator under native bridge
Change-Id: I24499cd22095e3c0002bc426ff4c9d64dfb9b8e6
(cherry picked from commit ada6248a1f195ed278e2956602f2f161d1baed6c)

Co-authored-by: dimitry <dimitry@google.com>