Make Day Explorer blames available in the Perfetto UI. (#1513)
Since they can take a while to calculate we hide them behind a command.

To make it more obvious when the command has finished, we open the first
group by default.

While we're here, clean up a little bit: import correct objects for
iterator types, and check for google3 with a fallthrough only once
instead of repeatedly.

Bug: http://b/417417058