Add handling for irq_vectors/local_timer_[entry|exit] events (#1493)
On x86, the local apic is usually what generates timer irqs.
Unfortunately, it's "special", so it doesn't register via the standard
generic irq code.

Instead, it uses irq_vectors/ which is an x86 specific trace event
system.

While one can capture irq_vectors/* and it shows up in the ftrace
tracks, it would be nice if the irq_vectors/* entries would be rendered
similarly to the IRQ track in the Perfetto UI.

So this pull request implements this feature. Adding the events to the
events_list and parsing the events, adding them to the IRQ track.

Bug: 415090521

---------

Signed-off-by: John Stultz <jstultz@google.com>