Fix crash when selecting first thread state slice
The thread state details panel loads the next and previous thread
slices and shows information about them, however these are not
guaranteed to exist (e.g. for first and last thread state slices).

Typescript should have warned us about this, but indexing into an
array in typescript by default doesn't just returns `T` instead of
`T | undefined`. We can enforce with `noUncheckedIndexedAccess` but
this causes a lot of errors.

For now, this patch just fixes the error at the source.

Change-Id: I8b777b4f489baea41b65de33e8ec6d7e16f56eb7