perfetto/ftrace: Add support for ftrace's max_graph_depth option (#1499)
In the perfetto config, add a new function_graph_max_depth option, which
sets the max depth for the function_graph tracer to follow in its
traces.

This is helpful to limit the amount of data that function_graph traces
collect, while still providing insight as to what might be taking the
time.

Bug: 417357433

Signed-off-by: John Stultz <jstultz@google.com>
Co-authored-by: Ryan <rsavitski@google.com>