Fix crash when selecting first thread state slice on any given thread state track (#1550)
The thread state details panel loads the next and previous thread slices
and shows information about them, however these are not guaranteed to
exist (e.g. for first and last thread state slices).

Typescript should have warned us about this, but indexing into an array
in typescript by default doesn't just returns `T` instead of `T |
undefined`. We can enforce with `noUncheckedIndexedAccess` but this
causes a lot of errors.

For now, this patch just fixes the error at the source.