importer: Add HPROF format support (#1514)
importer: Add HPROF format support

Parses and imports ART HPROF (Heap PROFile) dumps into Trace Processor.

The following tables are populated from the hprof:
- heap_graph_class
- heap_graph_object
- heap_graph_reference

Note that the following are unavailable in hprofs:
1. class_kind: Stored unknown
2. pid: Stored 0

The code is split into three main logical parts:

- ArtHprofParser:
  * Manages trace blob ingestion (ChunkedTraceReader).
  * Orchestrates parsing and storage population.
- HeapGraphBuilder:
* Handles low-level HPROF record parsing (header, records, heap dumps).
  * Constructs an in-memory HeapGraph with raw class and object data.
- HeapGraphResolver:
  * Takes the HeapGraph and resolves references and field values.
  * Marks reachable objects from GC roots.
  * Calculates native memory sizes (e.g., sun.misc.Cleaner).

The data model for the ART heap is defined in art_hprof_model.h,
providing representations of the following parsed entities

- ClassDefinition
- Object
- Field
- Reference

Primitive fields and array values are parsed, including decoding
Java UTF16 strings as UTF8. This is only used for parsing the
native_size out of NativeAllocationRegistry and propagating
the size through Cleaner references.

In the future, the data  values can be added to the perfetto tables.