frozen_ftrace: Erase the trace data after read (#1511)
After read out all data and flush the writer from "linux.frozen_ftrace"
data source, make sure to clear trace data in the kernel.

Background of this change is the upstream kernel patch[1], which makes
the kernel to recover the trace data that is already read in the
previous boot. This actually expects the perfetto AOT use case, because
it reads out the kernel data periodically. Thus we can not read most of
the data in the persistent ring buffer next boot time.
However, after perfetto read out the recovered previous boot data via
"linux.frozen_ftrace" data source, the data will be
recovered too (unless it is reused.) This means perfetto will read the
same data again next boot time. To prevent it, clear the trace data
after read via frozen_ftrace data source.

This does not change anything from user's viewpoint, because on the
current kernel, the read data is gone. And on the kernel with the
patch[1], it is cleared explicitly.

[1]
https://lore.kernel.org/all/174709742769.1964496.18203163435305117893.stgit@mhiramat.tok.corp.google.com/

Change-Id: Id4a7852c493fc4ac635396314eceb9d52ac57732