ci: run small jobs on GitHub's pool rather than ours
My theory is that we can make the CI more responsive by moving
the small jobs that don't require our sandbox on GitHub's pool.
This is particularly important for analyze.yml, which is an
input for the autoscale: if that doesn't run, the autoscaler
won't scale up much as it sees only one job (it will only scale
up after analyze.yml has run). This should reduce queue times
when the CI is idle.