importer: Fix Class objects and parser cleanup
Denote class objects with java.lang.Class<name>.
This disambiguates class objects from objects of the same
class. This fixes some strange duplicates in the flame
graph and this is the convention used for heap graphs anyways.

Refactored ArtHprofParser to reduce parameter passing.

Next up, fix the size of these objects.
They are currently zero, but perhaps should be same size as the
sum of static fields?

Test: tools/diff_test_trace_processor.py out/android/trace_processor_shell --name-filter '.*ArtHprof.*'
Change-Id: I1372f68aa84b0fa810eeccdb26c47adf70472441