Decentralize minimap (#1626)
This patch adds a new minimap provider extension point allowing plugins
to take control of what is displayed in the minimap area. Only one
provider can take control at a time, so we use a priority number to
define who 'wins'.

As a result, this patch moves the two minimap sources (sched and slices)
out to their relevant plugins in order to make use of this new extension
point.

This patch also updates the underlying mechanism behind the sched
minimap calculation to use interval_intersect, for more accurate minimap
bucketing.