Cache async track depths before at creation time. (#1635)
Async track depths are calculated using `experimental_slice_layout`
which, when passed verbatim to BaseSliceTrack, can be very slow when
cross joining on the result of the mipmap table. Currently we avoid this
by allowing the cross join to operate directly on the slice table,
because depths are not required anyway, but this is a hack and makes the
code hard to reason about and leaves rough edges for users of
DatasetSliceTrack.

This patch makes it so that the track depths are calculated up front
when an async slice track is created. This is much more efficient that
trying to cross join on it even once, and only happens one on startup,
leaving the cross joins to be fast.

This also means we can remove the getSqlJoinSource() hack in
BaseSliceTrack, which has been removed in this CL.