DatasetSliceTrack - Dynamic Datasets (#1638)
This change modifies the interface to `DatasetSliceTrack`, allowing it
to take either a static dataset or a dynamic dataset (a function that
returns a dataset, called on the track's onUpdate() hook).

This allows the dataset to be tweaked by the plugin that registers the
track depending on some other internal state of the plugin.

If the dataset's resultant query changes from frame to frame, the
track's internal state and any tables are dropped, and the track data is
reloaded.