Add bar chart support to the aggregation adapater (#1655)
In the past, thread state aggregation tabs used to have a little bar
chart at the top of the panel showing the proportions of thread states
within the selected area. This was lost after a refactor.

The way this used to work was very hacky - there was a special case for
the thread state tab that the core knew about.

This patch changes this special case to a more generic 'bar chart'
interface, where aggregators can specify bar chart data, which the
thread state aggregator makes use of to bring its bar chart back to
life.

Removing this special case also means that we can remove the hard
dependency on thread state structures and colors from the core, allowing
us to remove anything thread state related into the 'Sched' plugin.