Android Java SDK: use the same list to keep native objects in the root event builder. (#1668)
When building perfetto event, we need to keep references to the
'PerfettoPointer' objects that are used by that event. Otherwise, the GC
can collect them, and the 'native_delete' function will be called while
the native part of the object is still in use.

In this CL we fix the way we keep such references: we pass the same
'ArrayList' for all the child 'PerfettoTrackEventBuilder's, the same way
we pass 'ObjectsPool's, 'ObjectsCache's and other caches.

Tested:
```
tools/bazel test //:src_android_sdk_java_test_perfetto_trace_instrumentation_test
```