Handle CREATED, DEAD, and DESTROYED events in GenericTaskStateEvents
The thread table is populated with a start_ts when a CREATED event is
received for a non-existing thread. If any other state is received for a
non-existing thread, then the start_ts is left NULL (since we don't know
when the thread was actually created).

Now, threads are only ended when a DEAD event is received. The DESTROYED
event is only considered if its received before a DEAD event. If no utid
exists associated to a DESTROYED event, then the event is simply
ignored.

Also, we introduce a new error stat: generic_task_state_invalid_order,
to cover error cases in which events are received in an inproper
sequence. For example, the error stat is incremented if a CREATED event
is received for an already existing thread or if a DEAD event is
received after a DESTROYED event for the same thread.

Bug: 420976270