tp: dataframe: rework how sorting works in dataframes (#1706)
The existing implementation of sorting in dataframes was badly broken
for sparse null columns as it did not correctly compute the storage
index, instead just using the table index. This obviously caused sigsevs
when running the diff tests.

Turns out the old way of doing this could not possibly work and, as
we've known from existing implementation of sorting, doing stable_sort
on columns one by one in reverse order is actually a very bad way of
doing sorting.

Instead switch to a "row layout" and use the encoding scheme given by
https://arrow.apache.org/blog/2022/11/07/multi-column-sorts-in-arrow-rust-part-2/
and perform sorts based on binary bitstrings using memcmp instead. One
huge simplification we have over the quite complex code given by that
blog post is that we *do not* need to deal with variable sized
bitstrings. Because of our string pool, we can instead replace the
string with a "rank", computed by sorting all the referenced strings and
taking the position. By doing this, we significantly simplify both the
encoding scheme and the bytecode handling sorting logic.