probes: don't assume boot clock when preserve_ftrace_buffer is set
In case preserve_ftrace_buffer is set on Linux and the user has not
changed the clock to boot manually, the default clock will be "local".
But we were not running the code which reads the clock and snapshots the
value into the trace for clock sync.

Fix this by changing muxer to actually read and set the clock
unconditionally.