IME proto: add prevent_ime_startup_unless_text_editor
config_preventImeStartupUnlessTextEditor is an Android configuration
option that, when enabled, prevents the InputMethodManagerService from
starting up the input method (keyboard) unless a text editor is focused.

This configuration is used in many CtsInputMethodTestCases tests (like
CtsInputMethodTestCases:KeyboardVisibilityControlTest). Understanding
its value is essential for effective test debugging.

This CL is a port of ag/33261804 into external/perfetto.

Change-Id: I262743617ed9f50f1e762d99bd7faca04f7b7762