Trace Redaction - Expose "process tree adding duplicate entries" issue
When redacting a process trees, there are two action:

1. Remove all processes and threads that are not connected to the
   target package.

2. Add entries for synthetic threads and processes.

Adding synth thread and process entries must happen after removing
disconnected threads and processes to avoid removing the new
entries.

In this change, we split the two operations into their own
primitives to better expose a problem.

Filtering is simple, it optionally copies entries from one trace to
the other. This prevents the need for any special logic to respect
the trace's structure.

When adding entries, there needs to be additional logic. There is no
logic when adding synth threads/processes, so it adds them to every
process tree; it should only add them to one.

This change:

 1. Splits the two primitives, making it easier to see how the
    logic differs.

 2. Updates the tests to expose the issue and documents where the
    test is fudging values in order to pass.

A later change will be made to correct these issues.

Bug: 400093607