Remove redundant Track.title and rename TrackNode.title -> name
This patch removes the redundant 'title' property on the Track objects.
It's redundant because the string that's actually displayed on the
a track in the UI is the title property on the TrackNode, not the
Track, so the Track.title is basically unused. All it does it force
plugin developers to define this value twice.

Rationale: Why use TrackNode.title instead of Track.title?

Track 'titles' typically only make sense within the context of the
track's containing groups. For example the 'Actual Frames' track titles
make little sense outside the context of their containing process
group. Since groups are in the domain of TrackNodes, it makes sense
to locate this property here.

This patch also renames TrackNode.title -> name, as 'track name' is
a much more natural name for this concept.

Change-Id: I1ebd2435e341319201a86bdd558adfb06ad6d66e