tp: add support for Dataframe.SetCellUnchecked for SparseNullable cols (#1743)
While I thought we could avoid this, while working on migrating static
tables to Dataframe, it seems like there are too many parts of the
codebase which rely on being able to set sparse null values. We cannot
make them all dense null as that would bloat the memory use of TP
tremendously.

Therefore, add support for sparse null set calls. We'll mark these as
"slow" in the static API to make it clear to people this is a O(n)
funciton in the worst case.