Trace Redaction - move tgid out of tids in synth process
Before, we included a process's tgid in its tid list. The tgid was
also the pid. This required other logic to understand this and skip
the first thread. This made the code harder to understand.

This change moves the tgid out of the tids list and into a tgid field.
This ensures a synth process always has a pid/tgid now because its
needed in the constructor.

Bug: 400093607