Trace Redaction - Append fields
In order to copy the processes and threads from the old process tree
to the new process tree, we were decoding the old process tree and
re-encoding the process trees without any modification. This was
wasteful.

We could not copy the whole process tree because it would overwrite
a previous process tree. However we can copy the internals of the
message.

This change introduces an additional proto util function that wraps
the copy-base function and documents how we're using it.

Bug 400093607

Rather than decoding and appending fields, we can append the contents
of the source message.