tp: reduce performance cost of duplicate checking (#1809)
Doing a hashset add on every call was just too bad in terms of
performance. I tried using flathashmap and even wrote a custom optimized
hashset implementation and nothing worked.

In practice, the main case we care about duplicates is to build the
special side data structure for ids. Because of this, switch to using a
bitvector based implementation which is much more efficient.

On a represenative query, this takes the time taken to build the table
from 2.1s to 1.3s with the duplicate checking cost being ~zero