tp: clarify behaviour around dimension uniqueness and add option
This CL changes the documentation around dimension uniqueness for a set
of dimensions. Before we said we were expecting uniqueness but we were
not enforcing it. After discussions with folks internally, enforicng
uniqueness by default is probably not the right move as it prevents
aggregation across traces.

Instead, let's make the default *not* unique and then add an option to
enforce uniquenes. As we do this, also add proper enforcement when this
option is set.