tp: refactor to reduce compile times when changing interpreter (#1839)
Note: this change looks scary but is actually just a pure refactoring
with
no implementation change at all.

The current setup of header includes means that every time the
interpreter is changed (e.g. adding a new bytecode), ~400 files in trace
processor are recompiled. This sucks for fast iteration or debugging
complex query bugs.

Instead, refactor the interpreter into two files: one file for the
declaration and one for the actual definition of the code. Importantly
the second file is also a header so needs to be directly included.

Thankfully, by refactoring how TypedCursor works, we can make it so
that very few people have to worry about this and all of them live
inside the dataframe module