stdlib: Add table filtering for all IRQ slices
Create tables for exclusively hard IRQs, exclusively soft IRQs, and a
combined table of all IRQs.

The hard IRQ table and soft IRQ tables are almost symmetrical, with the
only difference being that the soft IRQ table does not have a parent_id,
since Linux guarantees that a soft IRQ will not preempt another IRQ, and
thus won't have an IRQ ancestor. The combined table of all IRQs is
simply the union of the hard IRQ and soft IRQ table.

Test: tools/diff_test_trace_processor.py out/linux_clang_release/trace_processor_shell --name-filter '.*linux_irqs'
Bug: 422478917
Signed-off-by: Samuel Wu <wusamuel@google.com>