Add usb group to predefined tracepoints
Create a usb group for DWC3 ftrace events.
DWC3 USB controller is a widely adopted SNPS IP, with linux upstream
support and extensitve use by companies like QCOM, Samsung and Google.
This would be helpful in debugging usb issues and analyzing usb
performance.
The counterpart in atrace is in ag/34015648.

Bug: 396043704
Test: out/android/perfetto_unittests
Test: capture and validate the trace
Signed-off-by: Roy Luo <royluo@google.com>