Show visual feedback while aggreagtions are loading (#1851)
Currently, when making a selection and the aggregation tabs take a long
time to load, nothing shows up in the details panel and we show a
message saying 'No details available for selection'. This is misleading
as usually there are some aggregations available but they are simply
loading, with no visual feedback.

This patch changes how the aggregation adapter works, so that it shows
all the relevant aggregation tabs immediately, and shows a loading
spinner while the aggregator is loading.


![image](https://github.com/user-attachments/assets/1c920088-88af-4195-87f3-8c06eaca8130)