Add a priority boost support to Perfetto.
This CL adds a feature to conditionally change the priority
of the `traced` process when the trace is running.
This feature is only supported on Android and Linux and requires
the `traced` to have 'CAP_SYS_NICE' capability (be started as root).

This is the first part of a bigger change.

Bug: b/423561591
Tested:
./perfetto_unittests