protos: tp: ui: add explicit enum for track_event merging
This CL adds an explicit proto for deciding how track_event tracks
should be merged together in the UI (and eventually the trace
processor). Until now we've relied on a very implicit merging
scheme but with this change, we can now properly define this
and allow users to choose how they want merging to happen.