ui: Allow tracks to have multiple kind tags
Currently tracks can only have one kind. This is limiting as we might
want to use a track like a slice track for the sake of searching, but
use it like an frames track for the sake of aggregation.

This patch switches out the singular kind tag with a kinds list of
tags. Now, tracks can have multiple kinds that bulk operations can use
to select tracks they operate on. This is very similar to how CSS
selectors and classes work.

This solves the concrete problem of allowing frame slice tracks to be
searched using dataset search.

Change-Id: I5cf0e293d6e23373103058529be49995649abd0a