Write a common function to parse the '/proc/self/stat' file.
The "comm" field may contain spaces, so splitting on a space is wrong.
In this implementation we search for '(' and ')' to extract the "comm"
field and split the rest of the string on space.

We'll switch the users to this new common function in the follow-up CLs.

See discussions here: https://news.ycombinator.com/item?id=34093845
and here (internal): go/bb-proc-stat

Tested:
./perfetto_unittests