tp: ftrace: allow for "local" ftrace clock (#1928)
There are cases (e.g. ftrace_config.preserve_ftrace_buffer) when ftrace
might
be timestamped using a non-boot clock, usually the default "local"
clock.

Allow such traces in trace_processor, doing a best-effort clock sync to
the
primary trace clock while treating the "local" clock as a
sequence-scoped
clock. As a simplification, we don't track the clock separately per cpu
as the
inter-cpu skew should be minimal for a cross-cpu trace to make sense in
the
first place.

This is based on, and supersedes #1825.