tprobes: rework ftrace clock snapshots (#1929)
After this patch:
* ftrace bundles reflect the actual value of "trace_clock", as we may
not be
setting it due to preserve_ftrace_buffer. Previously this would appear
as
  boot clock.
* Ftrace clock snapshots are supported when using secondary instances,
even
  though we don't expect any usage of this configuration at the moment.
* We continue treating "local" and "global" (which are technically
per-cpu
  clocks) as a single cross-cpu clock.

Implementation-wise, I've moved away from CpuReader retaining a pointer
to a
"last snapshot", and instead made each ReadCycle accept an explicit
snapshot to
use for all bundles. The latter makes the data dependency much clearer.

Based on, and supersedes #1824 + #1826.

Fixes: #1686