ui: Put AggregationPanel into Aggregator definition and refactor
This change moves the AggregationPenel (mostly used by Wattson) into
the aggregator definition itself, rather than getting injected in when
the tab is created. The aggregator is a natural place to put this as
it otherwise defines the aggregation.

To avoid some circular imports, this patch also moves all of the
aggregation-specific code into components rather than public, as it's
all related to the aggregation adapter which is a component anyway.

Change-Id: I3d5b0eeb8b898d489e904f4bcae7d014c0860fb0