ui: Add aggreagtion capabilities to DataGrid
This patch adds a new aggregation feature to the DataGrid component
that allows the caller to define an optional aggregates for each
column, choosing from a selection of aggregation functions:
- SUM
- AVG
- MIN
- MAX
- COUNT

Aggregations are displayed on the header bar underneath the column
title row, and all values are aligned with those underneath.

To save space, a symbol is used to indicate the aggregation function,
but the name of the function can be viewed by hovering the symbol.

Change-Id: I4b6c67df642c11542f00ace55bec8758a13710ae