ui: Put AggregationPanel into Aggregator definition and refactor (#1950)
This change moves the AggregationPanel (mostly used by Wattson) into the
aggregator definition itself, rather than getting injected in when the
tab is created. The aggregator is a natural place to put this as it
defines all the other properties of the aggregation, so it makes sense
that it also defines the panel component used.

To avoid some circular imports, this patch also moves all of the
aggregation-specific code into components rather than public, as it's
all related to the aggregation adapter which is a component.