ui: Add aggregation capabilities to DataGrid (#1953)
This patch adds a new aggregation feature to the DataGrid component that
allows the caller to define an optional aggregates for each column,
choosing from a selection of aggregation functions:
- SUM
- AVG
- MIN
- MAX
- COUNT

Aggregations are displayed on the header bar underneath the column title
row, and all values are aligned with those underneath.

To save space, a symbol is used to indicate the aggregation function,
but the name of the function can be viewed by hovering the symbol.

Example (from the widgets page):
<img width="527" alt="image"
src="https://github.com/user-attachments/assets/1c47db48-9a79-41d6-bf6c-067a4f26fdef"
/>