stdlib: Integrate IRQs as task for Wattson power attribution
In addition to threads/processes/packages, add IRQs as a potential task
for Wattson to attribute power for. This handles cases where tasks may
preempt another task, including for soft IRQs and hard IRQs.

For attributing idle transition costs, the first non-IRQ thread that
causes the wakeup is identified as the culprit. Without this, idle
transition costs would be exclusively attributed to IRQs, since the IRQs
are the mechanism that run and wakeup the CPU to process a task. For
"wakeup islands" that are purely IRQs processing, the IRQs are
attributed the wakeup power as if was a normal task.

Bug: 422478917
Signed-off-by: Samuel Wu <wusamuel@google.com>