Android.bp: generate jarjar-ed versions of Java SDK modules. (#1973)
In the framework, we don't want to depend on (and bundle) the
`dev.perfetto.sdk` classes.
Otherwise, when the 3P app uses Perfetto Java SDK classes with
the same name, but the different version, will be found both 
in App and Boot classpathes.

This problem is solved using jarjar: when building SDK for 
framework all classes
from the `dev.perfetto.sdk` package are moved to the new package.

Our JNI library resolves java classes by FQN name, so when building
for framework
the `-DPERFETTO_JNI_JARJAR_PREFIX=` compile flag should be
used to specify the new package.

We want both the 3P and the framework JNI lib to be buildable in GN,
so instead of generating the `_framework` version only for Android.bp
we add it as a normal GN target (not used to generate Bazel build
file, though).