ui: Add CPU PMU recording section
Add a new widget in the recording section to configure CPU PMUs to capture.
This module aims to simplify PMU selection based on Arm telemetry
specifications loaded by Perfetto.

The recording section allows users to load Arm telemetry specifications and
create multiple configurations. Each configuration lets the user specify which
CPU to record and which PMUs and sampling configuration to use.

In the UI, users can select PMUs to record by choosing metrics applicable to the
selected CPU. Two metric views are available: a flat list and a tree view that
follows the topdown specification for the chosen CPU.

Signed-off-by: Vincent Coubard <vincent.coubard@arm.com>