ui: Add option to connect to custom consumer socket
When updating probes, it is not always convenient to generate a new Android
build or replace the system Perfetto binaries.

This change introduces a way to connect to a tracebox instance running on the
target system, instead of the default Perfetto daemon.

When the feature flag `connectToLocalAbstractTraceboxSocket` is enabled, the
recording plugin connects to "@tracebox_traced_consumer" via a local abstract
socket rather than the standard system socket.

To use this option, start tracebox and websocket_bridge with
PERFETTO_CONSUMER_SOCK_NAME environment variable set to @tracebox_traced_consumer.

Signed-off-by: Vincent Coubard <vincent.coubard@arm.com>