Consider unique frame_id when calculating per-frame metrics. (#1956)
While extracting the frames in a CUJ, there are duplicate entries for a
CUJ. This is because currently there is no way to associate an actual
frame with a corresponding DrawFrame. At the metric level, this is an
immediate fix to make sure that the correct number of frames are used
for metric calculation.

Bug: 425562160
Test: tools/diff_test_trace_processor.py --name-filter
".*android_blocking_calls_cuj_per_frame.*"
out/linux_clang_debug/trace_processor_shell