tp: don't require single sequence emission for incremental counter tracks (#2002)
We were forcing incremental counter tracks to come from the same
sequence as incrementality does not make sense across sequences.
However, it's perfectly legit for two threads to have their own view of
the same counter: for example for thread timestamps.

This happens in practice when multiple libraries are independently using
the sdk and enabling chrome timestamps (e.g. webview + cronet).

Fix this by allowing for multiple sequences to contribute to the same
track with independent incremental values.