Add CPU filter into traced_perf
PMUs can generate large amounts of data which causes an overhead in term of
capture and processing. Perfetto capture the same PMUs for all of the CPUs
except if CPUs are selected individually.

This change introduce a parameter to restrict the capture on a specific
CPUIDs. The capture can be configured to restrict data captured to CPUs
that actually matters.

Signed-off-by: Vincent Coubard <vincent.coubard@arm.com>