Remove layer_id check for frames and CUJs.
Earlier implementation matched the layer_id for a frame with that of a
CUJ, to determine which frame should be attributed to which CUJ. The
per-frame blocking calls metric used this logic eventually. But this
causes blocking calls to be missed in certain cases, where there are
frames rendered during a CUJ, but for a different layer.

Bug: 429073039
Test: tools/diff_test_trace_processor.py --name-filter ".*android_blocking_calls_cuj_per_frame.*" out/linux_clang_debug/trace_processor_shell