ui: zoom in a fixed factor when focusing on instant events
This CL fixes the discrepency between how instant events and slice
events work when pressing 'f'. Specifically, in the case of instant
events, we just pan but this is not helpful as often people press f to
quickly zoom into a region of interest. Unfortunately as we don't have a
duration, we don't know "how much" we should zoom.

However, a fixed factor of 100x is a reasonable number looking at some
real traces and its better than not zooming at all.