base: speed up FlatHashMap::Clear() (#2030)
We were doing a full run through and rehash for no reason. In the end,
clear will want everything zeroed out and the tags empty. So just do
that directly.

This speeds up some pathalogical cases in trace processor where we were
spending a lot of time in the rehashing codepaths.