shared_lib: Don't call back if category was not enabled (#2094)
When a track_event data source instance is started, we call the category
callback only if the category is enabled for that config.

When a track event data source instance is stopped, we used to call the
category callback even if the category was disabled.

Let's only call it if the category was enabled for that config.