ui: Fix datagrid layout
Scroll the table not the toolbar, which means the paging controls stay put when scrolling large tables, and the table header stays pinned to the top.

Also fixed the widget layout to hide the overflow, limiting the size of widgets so that they get a chance to overflow and demonstrate scrolling.

Tested in the following places:
- Aggregation panels
- Query page
- Queries in the tab drawer
- Widgets page
- Explore page

Change-Id: I467a25ad2c061fca7dad1510d733cc87285f5f5a