ui: Fix datagrid layout (#2123)
Fixed the datagrid's css so that the table scrolls, not the toolbar when
space is limited, which means the toolbar / paging controls stay put and
the table header stays pinned to the top when scrolling large pages.

Also fixed the widget layout to hide the overflow, limiting the size of
widgets so that they get a chance to overflow and demonstrate scrolling.

Tested in the following places:
- Aggregation panels
- Query page
- Queries in the tab drawer
- Widgets page
- Explore page

Demo - notice the paging controls stay put while scrolling the table,
and the table header sticks to the top when scrolling vertically, but
says aligned with the columns when scrolling horizontally.


[e4c3f120-663d-4f3c-80f2-b79a45a1928f.webm](https://github.com/user-attachments/assets/30e445fe-4677-46a4-ab7e-c471d41e080e)