protos: add explicit enum for track_event merging (#1904)
This CL adds an explicit proto for deciding how track_event tracks
should be merged together in the UI (and eventually the trace
processor). Until now we've relied on a very implicit merging scheme but
with this change, we can now properly define this and allow users to
choose how they want merging to happen.

In the process, I've also had to put into writing how we handle track
names on TrackDescriptor: make the trace processor adhere to the proto
comments and add tests verifying this.