ui: Add ButtonGroup widget (#2131)
The ButtonGroup widget is a container widget housing a set of buttons
that are visually grouped together into one super-widget. The inside
borders are de-duplicated, and the inside rounded corners removed.

This is useful for when you have a set of radio buttons, or for adding a
drop down menu to an existing button.

Very similar to the SegmentedButtons widget, but offers more control
over the individual buttons.