ui: Add UI state sharing when trace loaded via postmessage
Traces loaded via postmessage are typically not sharable via permalink.
Instead, after the user clicks the share button, we can just show you
the URL to get back to that trace via the provider if a URL is supplied
via the postmessage args when the trace was first loaded. However this
doesn't allow for the ui state to be persisted.

This patch adds a feature where the ui state can be uploaded and the
hash injected back into the URL if a placeholder is provided, so that
the state may be restored when the trace is re-loaded.

Change-Id: I40d3114dc898f02edd0948d632dfa3bfebb9396f